ENH: add serializer class to extend ModelSerializer.
